refactor(core): Move all webhook related code out of the Workflow class
