+++ packages/workflow/src/NodeHelpers.ts
@@ -12,7 +12,6 @@ import { NodeConnectionType } from './Interfaces';
 import type {
 	FieldType,
 	IContextObject,
-	IHttpRequestMethods,
 	INode,
 	INodeCredentialDescription,
 	INodeIssueObjectProperty,
@@ -29,8 +28,6 @@ import type {
 	IParameterDependencies,
 	IRunExecutionData,
 	IVersionedNodeType,
-	IWebhookData,
-	IWorkflowExecuteAdditionalData,
 	NodeParameterValue,
 	ResourceMapperValue,
 	INodeTypeDescription,
@@ -1178,120 +1175,7 @@ export function getNodeParameters(
 }
 
 /**
- * Returns all the webhooks which should be created for the give node
- */
-export function getNodeWebhooks(
-	workflow: Workflow,
-	node: INode,
-	additionalData: IWorkflowExecuteAdditionalData,
-	ignoreRestartWebhooks = false,
-): IWebhookData[] {
-	if (node.disabled === true) {
-		// Node is disabled so webhooks will also not be enabled
-		return [];
-	}
-
-	const nodeType = workflow.nodeTypes.getByNameAndVersion(node.type, node.typeVersion);
-
-	if (nodeType.description.webhooks === undefined) {
-		// Node does not have any webhooks so return
-		return [];
-	}
-
-	const workflowId = workflow.id || '__UNSAVED__';
-	const mode = 'internal';
-
-	const returnData: IWebhookData[] = [];
-	for (const webhookDescription of nodeType.description.webhooks) {
-		if (ignoreRestartWebhooks && webhookDescription.restartWebhook === true) {
-			continue;
-		}
-
-		let nodeWebhookPath = workflow.expression.getSimpleParameterValue(
-			node,
-			webhookDescription.path,
-			mode,
-			{},
-		);
-		if (nodeWebhookPath === undefined) {
-			// TODO: Use a proper logger
-			console.error(
-				`No webhook path could be found for node "${node.name}" in workflow "${workflowId}".`,
-			);
-			continue;
-		}
-
-		nodeWebhookPath = nodeWebhookPath.toString();
-
-		if (nodeWebhookPath.startsWith('/')) {
-			nodeWebhookPath = nodeWebhookPath.slice(1);
-		}
-		if (nodeWebhookPath.endsWith('/')) {
-			nodeWebhookPath = nodeWebhookPath.slice(0, -1);
-		}
-
-		const isFullPath: boolean = workflow.expression.getSimpleParameterValue(
-			node,
-			webhookDescription.isFullPath,
-			'internal',
-			{},
-			undefined,
-			false,
-		) as boolean;
-		const restartWebhook: boolean = workflow.expression.getSimpleParameterValue(
-			node,
-			webhookDescription.restartWebhook,
-			'internal',
-			{},
-			undefined,
-			false,
-		) as boolean;
-		const path = getNodeWebhookPath(workflowId, node, nodeWebhookPath, isFullPath, restartWebhook);
-
-		const webhookMethods = workflow.expression.getSimpleParameterValue(
-			node,
-			webhookDescription.httpMethod,
-			mode,
-			{},
-			undefined,
-			'GET',
-		);
-
-		if (webhookMethods === undefined) {
-			// TODO: Use a proper logger
-			console.error(
-				`The webhook "${path}" for node "${node.name}" in workflow "${workflowId}" could not be added because the httpMethod is not defined.`,
-			);
-			continue;
-		}
-
-		let webhookId: string | undefined;
-		if ((path.startsWith(':') || path.includes('/:')) && node.webhookId) {
-			webhookId = node.webhookId;
-		}
-
-		String(webhookMethods)
-			.split(',')
-			.forEach((httpMethod) => {
-				if (!httpMethod) return;
-				returnData.push({
-					httpMethod: httpMethod.trim() as IHttpRequestMethods,
-					node: node.name,
-					path,
-					webhookDescription,
-					workflowId,
-					workflowExecuteAdditionalData: additionalData,
-					webhookId,
-				});
-			});
-	}
-
-	return returnData;
-}
-
-/**
  * Returns the webhook path
- *
  */
 export function getNodeWebhookPath(
 	workflowId: string,
@@ -1317,7 +1201,6 @@ export function getNodeWebhookPath(
 
 /**
  * Returns the webhook URL
- *
  */
 export function getNodeWebhookUrl(
 	baseUrl: string,
